feat(vs2022_third): C4244 fix double→float conversions in StressLineSectionSettingDlg.cpp — 1 file
Fix 47 C4244 warnings (double→float narrowing conversions) by adding
static_cast<float>() at assignment sites.

No behavior changes, only explicit casts to match target types.

+++ SRC/NanoSpec/StressLineSectionSettingDlg.cpp
@@ -1048,7 +1048,7 @@ void CStressLineSectionSettingDlg::OnPaint()
 
 	if((dStageSizeX / dStageSizeX) > ((StageRectRight - StageRectLeft) / (StageRectBottom - StageRectTop))) {
 		lDrawStageSizeX =(StageRectRight - StageRectLeft);
-		lDrawStageSizeY = ((StageRectRight - StageRectLeft) * dStageSizeX / dStageSizeX);
+		lDrawStageSizeY = static_cast<float>(((StageRectRight - StageRectLeft) * dStageSizeX / dStageSizeX));
 
 		m_dDrawStartStagePosX = StageRectLeft;
 		m_dDrawStartStagePosY = StageRectTop + ((StageRectBottom - StageRectTop - lDrawStageSizeY) / 2);
@@ -1056,7 +1056,7 @@ void CStressLineSectionSettingDlg::OnPaint()
 		m_dDrawEndStagePosY = m_dDrawStartStagePosY + lDrawStageSizeY;
 	}
 	else{
-		lDrawStageSizeX = ((StageRectBottom - StageRectTop) * dStageSizeX / dStageSizeX);
+		lDrawStageSizeX = static_cast<float>(((StageRectBottom - StageRectTop) * dStageSizeX / dStageSizeX));
 		lDrawStageSizeY = (StageRectBottom - StageRectTop);
 
 		m_dDrawStartStagePosX = StageRectLeft + ((StageRectRight - StageRectLeft - lDrawStageSizeX) / 2);
@@ -2201,7 +2201,7 @@ BOOL CStressLineSectionSettingDlg::CheckPinData()
 		if (!CheckInputValueDouble(Item.strText, &tmpPin))
 			return ( FALSE );
 
-		tmpPinX = tmpPin * MICROMETRE;
+		tmpPinX = static_cast<float>(tmpPin * MICROMETRE);
 // 2009.09.15 bagus stress èCê≥ --}--
 
 		Item.mask = GVIF_TEXT;
@@ -2213,7 +2213,7 @@ BOOL CStressLineSectionSettingDlg::CheckPinData()
 		if (!CheckInputValueDouble(Item.strText, &tmpPin))
 			return ( FALSE );
 
-		tmpPinY = tmpPin * MICROMETRE;
+		tmpPinY = static_cast<float>(tmpPin * MICROMETRE);
 // 2009.09.15 bagus stress èCê≥ --}--
 
 // 2009.09.08 bagus stress å¥ì_à íuëŒâû --{--
@@ -2411,7 +2411,7 @@ void CStressLineSectionSettingDlg::SetDefaultNecessaryThicknessMeasurementData()
 // 2009.09.11 bagus stress èCê≥ --}--
 	if (PinOrgY == 0) {
 		PinOrgY += m_StressConfig.lInvalidRangeEdge;
-		PinY = PinOrgY;
+		PinY = static_cast<float>(PinOrgY);
 	} else if (PinOrgY < 0) {
 		PinOrgY += m_StressConfig.lInvalidRangeEdge;
 		PinY = (long)((PinOrgY * 1000 - 500) / 1000);
@@ -2434,7 +2434,7 @@ void CStressLineSectionSettingDlg::SetDefaultNecessaryThicknessMeasurementData()
 //		PinX = (long)((((m_StressConfig.Size.dx / 2) - m_StressConfig.lInvalidRangeEdge) * 1000 + 500) / 1000);
 //		PinX *= -1;
 		if (PinOrgX == 0) {
-			PinX = PinOrgX;
+			PinX = static_cast<float>(PinOrgX);
 		} else if (PinOrgX < 0) {
 			PinX = (long)((PinOrgX * 1000 - 500) / 1000);
 		} else {
@@ -2446,7 +2446,7 @@ void CStressLineSectionSettingDlg::SetDefaultNecessaryThicknessMeasurementData()
 //			m_StressConfig.Line[i].MeasPos[j].lX = PinX;
 //			m_StressConfig.Line[i].MeasPos[j].lY = PinY;
 			if (j== PIN_INTERVAL-1) {
-				m_StressConfig.Line[i].MeasPos[j].lX = PinMaxX;
+				m_StressConfig.Line[i].MeasPos[j].lX = static_cast<float>(PinMaxX);
 //				m_StressConfig.Line[i].MeasPos[j].lY = PinMaxY;
 			} else {
 				m_StressConfig.Line[i].MeasPos[j].lX = PinX;
@@ -2672,32 +2672,32 @@ void CStressLineSectionSettingDlg::SamplePointGraph_DataSet()
 			switch (m_StageConfig.Dir.X) {
 			case 1:		// X:Å®
 			default:
-				lPinX = dOffSetValueX + m_StressConfig.Line[i].PinPos[j].lX;
+				lPinX = static_cast<float>(dOffSetValueX + m_StressConfig.Line[i].PinPos[j].lX);
 				break;
 			case 2:		// X:Å™
-				lPinY = dOffSetValueX - m_StressConfig.Line[i].PinPos[j].lX;
+				lPinY = static_cast<float>(dOffSetValueX - m_StressConfig.Line[i].PinPos[j].lX);
 				break;
 			case 3:		// X:Å©
-				lPinX = dOffSetValueX - m_StressConfig.Line[i].PinPos[j].lX;
+				lPinX = static_cast<float>(dOffSetValueX - m_StressConfig.Line[i].PinPos[j].lX);
 				break;
 			case 4:		// X:Å´
-				lPinY = dOffSetValueX + m_StressConfig.Line[i].PinPos[j].lX;
+				lPinY = static_cast<float>(dOffSetValueX + m_StressConfig.Line[i].PinPos[j].lX);
 				break;
 			}
 
 			switch (m_StageConfig.Dir.Y) {
 			case 1:		// Y:Å®
-				lPinX = dOffSetValueY + m_StressConfig.Line[i].PinPos[j].lY;
+				lPinX = static_cast<float>(dOffSetValueY + m_StressConfig.Line[i].PinPos[j].lY);
 				break;
 			case 2:		// Y:Å™
 			default:
-				lPinY = dOffSetValueY - m_StressConfig.Line[i].PinPos[j].lY;
+				lPinY = static_cast<float>(dOffSetValueY - m_StressConfig.Line[i].PinPos[j].lY);
 				break;
 			case 3:		// Y:Å©
-				lPinX = dOffSetValueY - m_StressConfig.Line[i].PinPos[j].lY;
+				lPinX = static_cast<float>(dOffSetValueY - m_StressConfig.Line[i].PinPos[j].lY);
 				break;
 			case 4:		// Y:Å´
-				lPinY = dOffSetValueY + m_StressConfig.Line[i].PinPos[j].lY;
+				lPinY = static_cast<float>(dOffSetValueY + m_StressConfig.Line[i].PinPos[j].lY);
 				break;
 			}
 			m_SamplePointGraph.AddPinData(lPinX,lPinY);
@@ -2727,40 +2727,40 @@ void CStressLineSectionSettingDlg::SamplePointGraph_DataSet()
 			switch (m_StageConfig.Dir.X) {
 			case 1:		// X:Å®
 			default:
-				tLineData.StartX = dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanStartPosX;
-				tLineData.EndX = dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanEndPosX;
+				tLineData.StartX = static_cast<float>(dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanStartPosX);
+				tLineData.EndX = static_cast<float>(dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanEndPosX);
 				break;
 			case 2:		// X:Å™
-				tLineData.StartY = dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanStartPosX;
-				tLineData.EndY = dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanEndPosX;
+				tLineData.StartY = static_cast<float>(dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanStartPosX);
+				tLineData.EndY = static_cast<float>(dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanEndPosX);
 				break;
 			case 3:		// X:Å©
-				tLineData.StartX = dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanStartPosX;
-				tLineData.EndX = dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanEndPosX;
+				tLineData.StartX = static_cast<float>(dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanStartPosX);
+				tLineData.EndX = static_cast<float>(dOffSetValueX - m_StressConfig.Line[i].SectPos[j].lScanEndPosX);
 				break;
 			case 4:		// X:Å´
-				tLineData.StartY = dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanStartPosX;
-				tLineData.EndY = dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanEndPosX;
+				tLineData.StartY = static_cast<float>(dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanStartPosX);
+				tLineData.EndY = static_cast<float>(dOffSetValueX + m_StressConfig.Line[i].SectPos[j].lScanEndPosX);
 				break;
 			}
 
 			switch (m_StageConfig.Dir.Y) {
 			case 1:		// Y:Å®
-				tLineData.StartX = dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY;
-				tLineData.EndX = dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY;
+				tLineData.StartX = static_cast<float>(dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY);
+				tLineData.EndX = static_cast<float>(dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY);
 				break;
 			case 2:		// Y:Å™
 			default:
-				tLineData.StartY = dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY;
-				tLineData.EndY = dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY;
+				tLineData.StartY = static_cast<float>(dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY);
+				tLineData.EndY = static_cast<float>(dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY);
 				break;
 			case 3:		// Y:Å©
-				tLineData.StartX = dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY;
-				tLineData.EndX = dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY;
+				tLineData.StartX = static_cast<float>(dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY);
+				tLineData.EndX = static_cast<float>(dOffSetValueY - m_StressConfig.Line[i].SectPos[j].lScanPosY);
 				break;
 			case 4:		// Y:Å´
-				tLineData.StartY = dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY;
-				tLineData.EndY = dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY;
+				tLineData.StartY = static_cast<float>(dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY);
+				tLineData.EndY = static_cast<float>(dOffSetValueY + m_StressConfig.Line[i].SectPos[j].lScanPosY);
 				break;
 			}
 // 2009.09.07 bagus stress å¥ì_à íuëŒâû --}--
@@ -2968,7 +2968,7 @@ void CStressLineSectionSettingDlg::DrawSample(CDC* pDC, int iOrg, int DirX, int 
 			Sampleplacement.rcNormalPosition.left	= Stageplacement.rcNormalPosition.left + dleft;
 			Sampleplacement.rcNormalPosition.right	= Stageplacement.rcNormalPosition.left + dright;
 			Sampleplacement.rcNormalPosition.top	= Stageplacement.rcNormalPosition.top  + dtop;
-			Sampleplacement.rcNormalPosition.bottom = Stageplacement.rcNormalPosition.top  + dbottom;
+			Sampleplacement.rcNormalPosition.bottom = static_cast<float>(Stageplacement.rcNormalPosition.top  + dbottom);
 			//éläpSampleï`âÊ
 			m_SamplePointGraph.SetWindowPlacement(&Sampleplacement);
 			m_SamplePointGraph.ShowWindow(SW_SHOW);
@@ -3742,13 +3742,13 @@ BOOL CStressLineSectionSettingDlg::SetAllInputData()
 		Item.row = i;
 		Item.col = 1;
 		m_PinListGrid.GetItem(&Item);
-		tmpPinX = strtod(Item.strText, NULL) * MICROMETRE;		// íPà ïœä∑(um->mm)
+		tmpPinX = static_cast<float>(strtod(Item.strText, NULL) * MICROMETRE);		// íPà ïœä∑(um->mm)
 
 		Item.mask = GVIF_TEXT;
 		Item.row = i;
 		Item.col = 2;
 		m_PinListGrid.GetItem(&Item);
-		tmpPinY = strtod(Item.strText, NULL) * MICROMETRE;		// íPà ïœä∑(um->mm)
+		tmpPinY = static_cast<float>(strtod(Item.strText, NULL) * MICROMETRE);		// íPà ïœä∑(um->mm)
 
 		PinX[i - PinFixRows] = tmpPinX;
 		PinY[i - PinFixRows] = tmpPinY;
